+++ R/utils-head.R
@@ -1,0 +1,10 @@
+#' head
+#'
+#' See \code{utils::\link[utils]{head}} for details.
+#'
+#' @name head
+#' @rdname head
+#' @keywords internal
+#' @export
+#' @importFrom utils head
+NULL