+++ R/utils-show_query.R
@@ -1,0 +1,10 @@
+#' show_query
+#'
+#' See \code{dplyr::\link[dplyr:explain]{show_query}} for details.
+#'
+#' @name show_query
+#' @rdname show_query
+#' @keywords internal
+#' @export
+#' @importFrom dplyr show_query
+NULL