+++ tests/testthat/test-List.R
@@ -1,9 +1,7 @@
-VicmapClient <- newClient()
-
 test_that("listLayers works", {
-  expect_equal(class(listLayers(VicmapClient)), "data.frame")
+  expect_equal(class(listLayers()), "data.frame")
 })
 
 test_that("listLayers filter works", {
-  expect_lte(nrow(listLayers(VicmapClient, pattern = stringr::regex("trees", ignore_case = T))), 5)
+  expect_lte(nrow(listLayers(pattern = stringr::regex("trees", ignore_case = T))), 5)
 })